Se corrige estructura
