+++ src/main/java/Boomberman/EstadoBomberman.java
@@ -1,3 +1,5 @@
+package Boomberman;
+
 public enum EstadoBomberman {
     ALIVE, DEAD
 }
