+++ src/main/java/Juego/Tablero/EstadoCelda.java
@@ -1,3 +1,5 @@
+package Juego.Tablero;
+
 public enum EstadoCelda {
     VACIA, PARED_MELAMINA, PARED_ACERO, ENEMIGO, BOMBERMAN, BOMBA
 }
