+++ src/main/java/Juego/Tablero/Posicion.java
@@ -1,7 +1,9 @@
+package Juego.Tablero;
+
 public class Posicion {
 
-    protected int x;
-    protected int y;
+    public int x;
+    public int y;
 
     public Posicion(int x, int y) {
         this.x = x;
