+++ src/main/java/Objetos/Enemigo.java
@@ -1,4 +1,10 @@
-public class ParedAcero implements Objeto{
+package Objetos;
+
+import Boomberman.Bomberman;
+import Juego.Tablero.Celda;
+import Juego.Tablero.EstadoCelda;
+
+public class Enemigo implements Objeto {
     @Override
     public void explotar(Celda celda) {
         celda.setObjeto(new Vacio());
@@ -6,11 +12,11 @@ public class ParedAcero implements Objeto{
 
     @Override
     public void moverA(Bomberman bomberman) {
-        bomberman.moverParedAcero();
+        bomberman.morir();
     }
 
     @Override
     public EstadoCelda estado() {
-        return EstadoCelda.PARED_ACERO;
+        return EstadoCelda.ENEMIGO;
     }
 }
