+++ src/main/java/Objetos/Objeto.java
@@ -1,3 +1,9 @@
+package Objetos;
+
+import Boomberman.Bomberman;
+import Juego.Tablero.Celda;
+import Juego.Tablero.EstadoCelda;
+
 public interface Objeto {
     public void explotar(Celda celda);
     public void moverA(Bomberman bomberman);
