+++ src/main/java/Objetos/Vacio.java
@@ -1,8 +1,13 @@
-public class Vacio implements Objeto {
-    @Override
-    public void explotar(Celda celda) {
+package Objetos;
 
-    }
+import Boomberman.Bomberman;
+import Juego.Tablero.Celda;
+import Juego.Tablero.EstadoCelda;
+
+public class Vacio implements Objeto {
+
+    @Override
+    public void explotar(Celda celda) {}
 
     @Override
     public void moverA(Bomberman bomberman) {
